avoid cast

+++ c_lines_bundle.c
@@ -93,7 +93,7 @@ int32 median(int32 *array, int32 length) {
     if ((length % 2) == 0) {
         return (array[(length/2) - 1] + array[length/2]) / 2.0;
     } else {
-        return array[(int)(length/2)];
+        return array[length/2];
     }
 }
